requirements.txt
flask==2.0.3
waitress==2.1.1
psycopg2==2.9.3
flask_migrate==2.5.3
pyjwt